refactor(rename): extract case-safe rename and link-rewrite into shared module
P01 of the uniform-rename plan. Move _rename_document_path and
_rewrite_incoming_refs (plus their helpers, regex, and budget const) out
of checks/structure.py into a new vaultcore/rename_ops.py so the structure
check and the upcoming rename backend share one implementation. The
structure check re-exports them under their old private names to keep its
test suite green unedited. Diagnostic types are imported lazily to avoid a
checks <-> rename_ops import cycle. Behavior is byte-for-byte unchanged.

Steps P01.S01-S04.
